Rename all components to use .jsx as file extension
