+++ client/components/admin/includes/AddBook.jsx
@@ -6,8 +6,6 @@ import { connect } from 'react-redux';
 import firebase from 'firebase';
 import ImageUploader from 'react-firebase-image-uploader';
 
-
-
 class AddBook extends Component {
   constructor(props) {
     super(props);
