+++ client/components/auth/login.jsx
@@ -27,12 +27,12 @@ class Login extends Component {
             [name]: value
         });
     }
-    handleFormSubmit(e){
+    handleFormSubmit(e) {
         e.preventDefault();
         getUserByEmail({ email: e.target.value })
-        .then((res) => {
-            console.log(res)
-        })
+            .then((res) => {
+                console.log(res)
+            })
     }
     handleSubmit(e) {
         e.preventDefault();
@@ -54,40 +54,40 @@ class Login extends Component {
     render() {
         const style = {
             button: {
-              backgroundColor: 'rgb(37, 76, 71)',
-              color: '#fff', float: 'right'
+                backgroundColor: 'rgb(37, 76, 71)',
+                color: '#fff', float: 'right'
             }
-          }
+        }
         return (
             <div id="login" className="col s12">
-                   <div id="forgot_password" className="modal">
-         <div className="modal-content">
-           <h4 style={{ alignContent: 'center' }}>Request For a new password</h4>
-           <div className="row">
-             <form name="forgot_pass" action='/search' className="col s12"
-               onSubmit={this.handleFormSubmit}>
-               <div className="add-book">
-                 <div className="row">
-                   <div className="input-field col s12">
-                     <input
-                       id="name"
-                       type="email"
-                       name="text"
-                       onChange={this.onChange}
-                       className="validate"
-                       required />
-                     <label htmlFor="isbn">Enter you Email</label>
-                   </div>
-                 </div>
-                 </div>
-               <button style={style.button}
-                 className="btn waves-effect waves-light"
-                 type="submit" name="submit">Search
+                <div id="forgot_password" className="modal">
+                    <div className="modal-content">
+                        <h4 style={{ alignContent: 'center' }}>Request For a new password</h4>
+                        <div className="row">
+                            <form name="forgot_pass" action='/search' className="col s12"
+                                onSubmit={this.handleFormSubmit}>
+                                <div className="add-book">
+                                    <div className="row">
+                                        <div className="input-field col s12">
+                                            <input
+                                                id="name"
+                                                type="email"
+                                                name="text"
+                                                onChange={this.onChange}
+                                                className="validate"
+                                                required />
+                                            <label htmlFor="isbn">Enter you Email</label>
+                                        </div>
+                                    </div>
+                                </div>
+                                <button style={style.button}
+                                    className="btn waves-effect waves-light"
+                                    type="submit" name="submit">Search
                      </button>
-             </form>
-           </div>
-         </div>
-       </div>
+                            </form>
+                        </div>
+                    </div>
+                </div>
                 <div style={{ color: 'red', textAlign: 'center' }}>{this.state.loginError}</div>
                 <form className="col s12" onSubmit={this.handleSubmit}>
                     <div className="form-container">
@@ -111,7 +111,7 @@ class Login extends Component {
                             <button className="btn waves-effect waves-light teal" type="submit" name="action" disabled={this.state.isLoading}>Login</button>
                             <br />
                             <br />
-                          
+
                             <a data-target="forgot_password" className="modal-trigger" href="#forgot_password">Forgotten password?</a>
                         </center>
                     </div>
